+++ src/test/java/com/google/cloud/spanner/pgadapter/EmulatedPsqlMockServerTest.java
@@ -26,6 +26,8 @@ import com.google.cloud.spanner.Statement;
 import com.google.cloud.spanner.pgadapter.error.PGException;
 import com.google.cloud.spanner.pgadapter.statements.IntermediateStatement;
 import com.google.common.collect.ImmutableList;
+import com.google.protobuf.ListValue;
+import com.google.protobuf.Value;
 import com.google.rpc.ResourceInfo;
 import com.google.spanner.admin.database.v1.Database;
 import com.google.spanner.admin.database.v1.DatabaseDialect;
@@ -37,6 +39,7 @@ import com.google.spanner.v1.DatabaseName;
 import com.google.spanner.v1.ExecuteBatchDmlRequest;
 import com.google.spanner.v1.ExecuteSqlRequest;
 import com.google.spanner.v1.SessionName;
+import com.google.spanner.v1.TypeCode;
 import io.grpc.Metadata;
 import io.grpc.Status;
 import io.grpc.StatusRuntimeException;
@@ -58,7 +61,6 @@ import org.postgresql.PGProperty;
 
 @RunWith(JUnit4.class)
 public class EmulatedPsqlMockServerTest extends AbstractMockServerTest {
-  // @Rule public Timeout globalTimeout = Timeout.seconds(10);
 
   private static final String INSERT1 = "insert into foo values (1)";
   private static final String INSERT2 = "insert into foo values (2)";
@@ -305,6 +307,51 @@ public class EmulatedPsqlMockServerTest extends AbstractMockServerTest {
       assertEquals(
           "ERROR: prepared statement my_prepared_statement does not exist",
           sqlException.getMessage());
+    }
+  }
+
+  @Test
+  public void testTimezone() throws SQLException {
+    String sql = "select ts from foo";
+    mockSpanner.putStatementResult(
+        StatementResult.query(
+            Statement.of(sql),
+            com.google.spanner.v1.ResultSet.newBuilder()
+                .setMetadata(createMetadata(ImmutableList.of(TypeCode.TIMESTAMP)))
+                .addRows(
+                    ListValue.newBuilder()
+                        .addValues(
+                            Value.newBuilder()
+                                .setStringValue("2023-01-06T11:49:15.123456789Z")
+                                .build())
+                        .build())
+                .build()));
+
+    try (Connection connection = DriverManager.getConnection(createUrl("my-db"))) {
+      connection.createStatement().execute("set time zone cet");
+      try (ResultSet resultSet = connection.createStatement().executeQuery(sql)) {
+        while (resultSet.next()) {
+          assertEquals("2023-01-06 12:49:15.123456+01", resultSet.getString(1));
+        }
+      }
+      connection.createStatement().execute("set time zone ist");
+      try (ResultSet resultSet = connection.createStatement().executeQuery(sql)) {
+        while (resultSet.next()) {
+          assertEquals("2023-01-06 17:19:15.123456+05:30", resultSet.getString(1));
+        }
+      }
+      connection.createStatement().execute("set time zone 'America/Los_Angeles'");
+      try (ResultSet resultSet = connection.createStatement().executeQuery(sql)) {
+        while (resultSet.next()) {
+          assertEquals("2023-01-06 03:49:15.123456-08", resultSet.getString(1));
+        }
+      }
+      connection.createStatement().execute("set time zone -12");
+      try (ResultSet resultSet = connection.createStatement().executeQuery(sql)) {
+        while (resultSet.next()) {
+          assertEquals("2023-01-05 23:49:15.123456-12", resultSet.getString(1));
+        }
+      }
     }
   }