+++ src/test/java/com/google/cloud/spanner/pgadapter/JdbcMockServerTest.java
@@ -42,6 +42,7 @@ import com.google.cloud.spanner.pgadapter.wireprotocol.ControlMessage.PreparedTy
 import com.google.cloud.spanner.pgadapter.wireprotocol.DescribeMessage;
 import com.google.cloud.spanner.pgadapter.wireprotocol.ExecuteMessage;
 import com.google.cloud.spanner.pgadapter.wireprotocol.ParseMessage;
+import com.google.common.base.Stopwatch;
 import com.google.common.collect.ImmutableList;
 import com.google.protobuf.ListValue;
 import com.google.protobuf.Value;
@@ -77,9 +78,11 @@ import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 import java.time.temporal.ChronoUnit;
 import java.util.*;
+import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import org.junit.BeforeClass;
+import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
@@ -507,22 +510,27 @@ public class JdbcMockServerTest extends AbstractMockServerTest {
             assertEquals(3.14d, resultSet.getDouble(++index), 0.0d);
             assertEquals(100, resultSet.getInt(++index));
             assertEquals(new BigDecimal("6.626"), resultSet.getBigDecimal(++index));
-            if (preparedThreshold < 0) {
-              // The binary format will truncate the timestamp value to microseconds.
-              assertEquals(
-                  truncatedOffsetDateTime, resultSet.getObject(++index, OffsetDateTime.class));
-            } else {
-              assertEquals(offsetDateTime, resultSet.getObject(++index, OffsetDateTime.class));
-            }
+            assertEquals(
+                truncatedOffsetDateTime, resultSet.getObject(++index, OffsetDateTime.class));
             assertEquals(LocalDate.of(2022, 3, 29), resultSet.getObject(++index, LocalDate.class));
             assertEquals("test", resultSet.getString(++index));
             assertEquals("{\"key\": \"value\"}", resultSet.getString(++index));
+
+            for (int col = 1; col <= resultSet.getMetaData().getColumnCount(); col++) {
+              assertNotNull(resultSet.getObject(col));
+            }
+
             assertFalse(resultSet.next());
           }
         }
       }
 
-      List<ExecuteSqlRequest> requests = mockSpanner.getRequestsOfType(ExecuteSqlRequest.class);
+      List<ExecuteSqlRequest> requests =
+          mockSpanner.getRequests().stream()
+              .filter(request -> request instanceof ExecuteSqlRequest)
+              .map(request -> (ExecuteSqlRequest) request)
+              .filter(request -> request.getSql().equals(pgSql))
+              .collect(Collectors.toList());
       // Prepare threshold less than 0 means use binary transfer + DESCRIBE statement.
       assertEquals(preparedThreshold < 0 ? 2 : 1, requests.size());
 
@@ -1761,7 +1769,9 @@ public class JdbcMockServerTest extends AbstractMockServerTest {
                 .bind("p9")
                 .to("test")
                 .bind("p10")
-                .to("{\"key\": \"value\"}")
+                // TODO: Change to jsonb when https://github.com/googleapis/java-spanner/pull/2182
+                //       has been merged.
+                .to(com.google.cloud.spanner.Value.json("{\"key\": \"value\"}"))
                 .build(),
             com.google.spanner.v1.ResultSet.newBuilder()
                 .setMetadata(ALL_TYPES_METADATA)
@@ -1784,8 +1794,8 @@ public class JdbcMockServerTest extends AbstractMockServerTest {
         assertEquals(Types.TIMESTAMP, parameterMetaData.getParameterType(7));
         assertEquals(Types.DATE, parameterMetaData.getParameterType(8));
         assertEquals(Types.VARCHAR, parameterMetaData.getParameterType(9));
-        // TODO: Enable when support for JSONB has been enabled.
-        // assertEquals(Types.OTHER, parameterMetaData.getParameterType(10));
+        assertEquals(Types.OTHER, parameterMetaData.getParameterType(10));
+
         ResultSetMetaData metadata = statement.getMetaData();
         assertEquals(10, metadata.getColumnCount());
         assertEquals(Types.BIGINT, metadata.getColumnType(1));
@@ -1797,8 +1807,7 @@ public class JdbcMockServerTest extends AbstractMockServerTest {
         assertEquals(Types.TIMESTAMP, metadata.getColumnType(7));
         assertEquals(Types.DATE, metadata.getColumnType(8));
         assertEquals(Types.VARCHAR, metadata.getColumnType(9));
-        // TODO: Enable when support for JSONB has been enabled.
-        // assertEquals(Types.OTHER, metadata.getColumnType(10));
+        assertEquals(Types.OTHER, metadata.getColumnType(10));
 
         int index = 0;
         statement.setLong(++index, 1L);
@@ -1810,7 +1819,7 @@ public class JdbcMockServerTest extends AbstractMockServerTest {
         statement.setObject(++index, zonedDateTime);
         statement.setObject(++index, LocalDate.of(2022, 3, 29));
         statement.setString(++index, "test");
-        statement.setString(++index, "{\"key\": \"value\"}");
+        statement.setObject(++index, createJdbcPgJsonbObject("{\"key\": \"value\"}"));
 
         try (ResultSet resultSet = statement.executeQuery()) {
           assertTrue(resultSet.next());
@@ -2794,7 +2803,7 @@ public class JdbcMockServerTest extends AbstractMockServerTest {
           }
           count++;
         }
-        assertEquals(358, count);
+        assertEquals(359, count);
       }
     }
   }
@@ -2954,6 +2963,36 @@ public class JdbcMockServerTest extends AbstractMockServerTest {
             createUrl() + "?options=-c%20server_version=5.2 custom version")) {
       verifySettingValue(connection, "server_version", "5.2 custom version");
       verifySettingValue(connection, "server_version_num", "50002");
+    }
+  }
+
+  @Test
+  public void testSetConnectionApiOptionInConnectionOptions() throws SQLException {
+    try (Connection connection =
+        DriverManager.getConnection(
+            createUrl() + "?options=-c%20spanner.autocommit_dml_mode='partitioned_non_atomic'")) {
+      verifySettingValue(connection, "spanner.autocommit_dml_mode", "PARTITIONED_NON_ATOMIC");
+    }
+  }
+
+  @Test
+  public void testSetInvalidConnectionApiOptionInConnectionOptions() throws SQLException {
+    try (Connection connection =
+        DriverManager.getConnection(
+            createUrl() + "?options=-c%20spanner.read_only_staleness='foo'")) {
+      verifySettingValue(connection, "spanner.read_only_staleness", "STRONG");
+    }
+  }
+
+  @Test
+  public void testSetInvalidAndValidConnectionApiOptionInConnectionOptions() throws SQLException {
+    try (Connection connection =
+        DriverManager.getConnection(
+            createUrl()
+                + "?options=-c%20spanner.read_only_staleness='foo'"
+                + "%20-c%20spanner.autocommit_dml_mode='partitioned_non_atomic'")) {
+      verifySettingValue(connection, "spanner.read_only_staleness", "STRONG");
+      verifySettingValue(connection, "spanner.autocommit_dml_mode", "PARTITIONED_NON_ATOMIC");
     }
   }
 
@@ -3242,6 +3281,117 @@ public class JdbcMockServerTest extends AbstractMockServerTest {
     }
 
     assertEquals(1, mockSpanner.countRequestsOfType(ExecuteSqlRequest.class));
+  }
+
+  @Test
+  public void testForceAutocommit() throws SQLException {
+    try (Connection connection = DriverManager.getConnection(createUrl())) {
+      try (java.sql.Statement statement = connection.createStatement()) {
+        // Force PGAdapter to use autocommit mode in all cases.
+        statement.execute("set spanner.force_autocommit to on");
+        // Set the connection to transactional mode. This will be ignored by PGAdapter.
+        connection.setAutoCommit(false);
+        statement.execute(SELECT1.getSql());
+        statement.execute(UPDATE_STATEMENT.getSql());
+        statement.execute(INSERT_STATEMENT.getSql());
+      }
+    }
+    assertEquals(3, mockSpanner.countRequestsOfType(ExecuteSqlRequest.class));
+    assertEquals(2, mockSpanner.countRequestsOfType(CommitRequest.class));
+    ExecuteSqlRequest selectRequest = mockSpanner.getRequestsOfType(ExecuteSqlRequest.class).get(0);
+    assertEquals(selectRequest.getSql(), SELECT1.getSql());
+    assertTrue(selectRequest.getTransaction().hasSingleUse());
+    assertTrue(selectRequest.getTransaction().getSingleUse().hasReadOnly());
+    ExecuteSqlRequest updateRequest = mockSpanner.getRequestsOfType(ExecuteSqlRequest.class).get(1);
+    assertTrue(updateRequest.getTransaction().hasBegin());
+    assertTrue(updateRequest.getTransaction().getBegin().hasReadWrite());
+    ExecuteSqlRequest insertRequest = mockSpanner.getRequestsOfType(ExecuteSqlRequest.class).get(2);
+    assertTrue(insertRequest.getTransaction().hasBegin());
+    assertTrue(insertRequest.getTransaction().getBegin().hasReadWrite());
+  }
+
+  @Test
+  public void testForceAutocommitWithPdml() throws SQLException {
+    try (Connection connection = DriverManager.getConnection(createUrl())) {
+      try (java.sql.Statement statement = connection.createStatement()) {
+        // Force PGAdapter to use autocommit mode in all cases.
+        statement.execute("set spanner.force_autocommit to on");
+        statement.execute("set spanner.autocommit_dml_mode to 'partitioned_non_atomic'");
+        // Set the connection to transactional mode. This will be ignored by PGAdapter.
+        connection.setAutoCommit(false);
+        statement.execute(UPDATE_STATEMENT.getSql());
+        statement.execute(INSERT_STATEMENT.getSql());
+      }
+    }
+    assertEquals(2, mockSpanner.countRequestsOfType(ExecuteSqlRequest.class));
+    assertEquals(1, mockSpanner.countRequestsOfType(CommitRequest.class));
+    // The update statement should use a partitioned DML transaction.
+    ExecuteSqlRequest updateRequest = mockSpanner.getRequestsOfType(ExecuteSqlRequest.class).get(0);
+    assertFalse(updateRequest.getTransaction().hasBegin());
+    assertTrue(updateRequest.getTransaction().hasId());
+    assertEquals(1, mockSpanner.countRequestsOfType(BeginTransactionRequest.class));
+    assertEquals(
+        1,
+        mockSpanner.getRequestsOfType(BeginTransactionRequest.class).stream()
+            .filter(request -> request.getOptions().hasPartitionedDml())
+            .count());
+    // The insert statement will use a normal read/write transaction.
+    ExecuteSqlRequest insertRequest = mockSpanner.getRequestsOfType(ExecuteSqlRequest.class).get(1);
+    assertTrue(insertRequest.getTransaction().hasBegin());
+    assertTrue(insertRequest.getTransaction().getBegin().hasReadWrite());
+  }
+
+  @Test
+  public void testInsertPdml() throws SQLException {
+    String sql = "insert into my_table (id, value) values (1, 'one')";
+    mockSpanner.putStatementResult(
+        StatementResult.exception(
+            Statement.of(sql),
+            Status.FAILED_PRECONDITION
+                .withDescription(
+                    "insert statements are not allowed in Partitioned DML transactions")
+                .asRuntimeException()));
+
+    try (Connection connection = DriverManager.getConnection(createUrl())) {
+      try (java.sql.Statement statement = connection.createStatement()) {
+        statement.execute("set spanner.autocommit_dml_mode to 'partitioned_non_atomic'");
+        SQLException exception = assertThrows(SQLException.class, () -> statement.execute(sql));
+        assertEquals(
+            "ERROR: insert statements are not allowed in Partitioned DML transactions",
+            exception.getMessage());
+      }
+    }
+    assertEquals(1, mockSpanner.countRequestsOfType(BeginTransactionRequest.class));
+    assertEquals(
+        1,
+        mockSpanner.getRequestsOfType(BeginTransactionRequest.class).stream()
+            .filter(request -> request.getOptions().hasPartitionedDml())
+            .count());
+    assertEquals(1, mockSpanner.countRequestsOfType(ExecuteSqlRequest.class));
+    assertEquals(0, mockSpanner.countRequestsOfType(CommitRequest.class));
+  }
+
+  @Ignore("Only used for manual performance testing")
+  @Test
+  public void testBasePerformance() throws SQLException {
+    final int numRuns = 1000;
+    String sql = "select * from random_benchmark";
+    RandomResultSetGenerator generator = new RandomResultSetGenerator(10, Dialect.POSTGRESQL);
+    for (int run = 0; run < numRuns; run++) {
+      mockSpanner.putStatementResult(
+          StatementResult.query(Statement.of(sql + run), generator.generate()));
+    }
+    try (Connection connection = DriverManager.getConnection(createUrl())) {
+      Stopwatch watch = Stopwatch.createStarted();
+      for (int run = 0; run < numRuns; run++) {
+        try (ResultSet resultSet = connection.createStatement().executeQuery(sql + run)) {
+          while (resultSet.next()) {
+            // ignore
+          }
+        }
+      }
+      System.out.printf("Elapsed: %dms\n", watch.elapsed(TimeUnit.MILLISECONDS));
+    }
   }
 
   private void verifySettingIsNull(Connection connection, String setting) throws SQLException {